API - add Get one method - java

+++ examples/example2/templates/java/Repository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ${SINGLE_ROUTE_NAME_CAPITALIZE}Repository extends JpaRepository<${SINGLE_ROUTE_NAME_CAPITALIZE}Entity, Long> {
+public interface ${SINGLE_ROUTE_NAME_CAPITALIZE}Repository extends JpaRepository<${SINGLE_ROUTE_NAME_CAPITALIZE}Entity, Integer> {
 }
